+++ test/unit/auth/token-generator.spec.ts
@@ -22,22 +22,21 @@ import * as chai from 'chai';
 import * as sinon from 'sinon';
 import * as sinonChai from 'sinon-chai';
 import * as chaiAsPromised from 'chai-as-promised';
+import * as nock from 'nock';
 
 import * as mocks from '../../resources/mocks';
-import {
-  FirebaseTokenGenerator, SESSION_COOKIE_INFO, ID_TOKEN_INFO,
-} from '../../../src/auth/token-generator';
-import * as verifier from '../../../src/auth/token-verifier';
-import {FirebaseAuthError, AuthClientErrorCode} from '../../../src/utils/error';
+import {FirebaseTokenGenerator, ServiceAccountSigner, IAMSigner} from '../../../src/auth/token-generator';
 
 import {Certificate} from '../../../src/auth/credential';
+import { AuthorizedHttpClient, HttpClient } from '../../../src/utils/api-request';
+import { FirebaseApp } from '../../../src/firebase-app';
+import * as utils from '../utils';
 
 chai.should();
 chai.use(sinonChai);
 chai.use(chaiAsPromised);
 
 const expect = chai.expect;
-
 
 const ALGORITHM = 'RS256';
 const ONE_HOUR_IN_SECONDS = 60 * 60;
@@ -46,7 +45,6 @@ const BLACKLISTED_CLAIMS = [
   'acr', 'amr', 'at_hash', 'aud', 'auth_time', 'azp', 'cnf', 'c_hash', 'exp', 'iat', 'iss', 'jti',
   'nbf', 'nonce',
 ];
-
 
 /**
  * Verifies a token is signed with the private key corresponding to the provided public key.
@@ -69,13 +67,203 @@ function verifyToken(token: string, publicKey: string): Promise<object> {
   });
 }
 
+describe('CryptoSigner', () => {
+  describe('ServiceAccountSigner', () => {
+    it('should throw given no arguments', () => {
+      expect(() => {
+        const anyServiceAccountSigner: any = ServiceAccountSigner;
+        return new anyServiceAccountSigner();
+      }).to.throw('Must provide a certificate to initialize ServiceAccountSigner');
+    });
+
+    it('should not throw given a valid certificate', () => {
+      expect(() => {
+        return new ServiceAccountSigner(new Certificate(mocks.certificateObject));
+      }).not.to.throw();
+    });
+
+    it('should sign using the private_key in the certificate', () => {
+      const payload = Buffer.from('test');
+      const cert = new Certificate(mocks.certificateObject);
+
+      const crypto = require('crypto');
+      const rsa = crypto.createSign('RSA-SHA256');
+      rsa.update(payload);
+      const result = rsa.sign(cert.privateKey, 'base64');
+
+      const signer = new ServiceAccountSigner(cert);
+      return signer.sign(payload).then((signature) => {
+        expect(signature.toString('base64')).to.equal(result);
+      });
+    });
+
+    it('should return the client_email from the certificate', () => {
+      const cert = new Certificate(mocks.certificateObject);
+      const signer = new ServiceAccountSigner(cert);
+      return signer.getAccountId().should.eventually.equal(cert.clientEmail);
+    });
+  });
+
+  describe('IAMSigner', () => {
+    let mockApp: FirebaseApp;
+    const mockAccessToken: string = utils.generateRandomAccessToken();
+
+    before(() => {
+      utils.mockFetchAccessTokenRequests(mockAccessToken);
+    });
+
+    after(() => nock.cleanAll());
+
+    beforeEach(() => {
+      mockApp = mocks.app();
+    });
+
+    afterEach(() => {
+      return mockApp.delete();
+    });
+
+    it('should throw given no arguments', () => {
+      expect(() => {
+        const anyIAMSigner: any = IAMSigner;
+        return new anyIAMSigner();
+      }).to.throw('Must provide a HTTP client to initialize IAMSigner');
+    });
+
+    describe('explicit service account ID', () => {
+      const response = {signature: Buffer.from('testsignature').toString('base64')};
+      const input = Buffer.from('input');
+      const signRequest = {
+        method: 'POST',
+        url: `https://iam.googleapis.com/v1/projects/-/serviceAccounts/test-service-account:signBlob`,
+        headers: {Authorization: `Bearer ${mockAccessToken}`},
+        data: {bytesToSign: input.toString('base64')},
+      };
+      let stub: sinon.SinonStub;
+
+      afterEach(() => {
+        stub.restore();
+      });
+
+      it('should sign using the IAM service', () => {
+        const expectedResult = utils.responseFrom(response);
+        stub = sinon.stub(HttpClient.prototype, 'send').resolves(expectedResult);
+        const requestHandler = new AuthorizedHttpClient(mockApp);
+        const signer = new IAMSigner(requestHandler, 'test-service-account');
+        return signer.sign(input).then((signature) => {
+          expect(signature.toString('base64')).to.equal(response.signature);
+          expect(stub).to.have.been.calledOnce.and.calledWith(signRequest);
+        });
+      });
+
+      it('should fail if the IAM service responds with an error', () => {
+        const expectedResult = utils.errorFrom({
+          error: {
+            status: 'PROJECT_NOT_FOUND',
+            message: 'test reason',
+          },
+        });
+        stub = sinon.stub(HttpClient.prototype, 'send').rejects(expectedResult);
+        const requestHandler = new AuthorizedHttpClient(mockApp);
+        const signer = new IAMSigner(requestHandler, 'test-service-account');
+        return signer.sign(input).catch((err) => {
+          const message = 'test reason; Please refer to ' +
+            'https://firebase.google.com/docs/auth/admin/create-custom-tokens for more details on ' +
+            'how to use and troubleshoot this feature.';
+          expect(err.message).to.equal(message);
+          expect(stub).to.have.been.calledOnce.and.calledWith(signRequest);
+        });
+      });
+
+      it('should return the explicitly specified service account', () => {
+        const signer = new IAMSigner(new AuthorizedHttpClient(mockApp), 'test-service-account');
+        return signer.getAccountId().should.eventually.equal('test-service-account');
+      });
+    });
+
+    describe('auto discovered service account', () => {
+      const input = Buffer.from('input');
+      const response = {signature: Buffer.from('testsignature').toString('base64')};
+      const metadataRequest = {
+        method: 'GET',
+        url: `http://metadata/computeMetadata/v1/instance/service-accounts/default/email`,
+        headers: {'Metadata-Flavor': 'Google'},
+      };
+      const signRequest = {
+        method: 'POST',
+        url: `https://iam.googleapis.com/v1/projects/-/serviceAccounts/discovered-service-account:signBlob`,
+        headers: {Authorization: `Bearer ${mockAccessToken}`},
+        data: {bytesToSign: input.toString('base64')},
+      };
+      let stub: sinon.SinonStub;
+
+      afterEach(() => {
+        stub.restore();
+      });
+
+      it('should sign using the IAM service', () => {
+        stub = sinon.stub(HttpClient.prototype, 'send');
+        stub.onCall(0).resolves(utils.responseFrom('discovered-service-account'));
+        stub.onCall(1).resolves(utils.responseFrom(response));
+        const requestHandler = new AuthorizedHttpClient(mockApp);
+        const signer = new IAMSigner(requestHandler);
+        return signer.sign(input).then((signature) => {
+          expect(signature.toString('base64')).to.equal(response.signature);
+          expect(stub).to.have.been.calledTwice;
+          expect(stub.getCall(0).args[0]).to.deep.equal(metadataRequest);
+          expect(stub.getCall(1).args[0]).to.deep.equal(signRequest);
+        });
+      });
+
+      it('should fail if the IAM service responds with an error', () => {
+        const expectedResult = {
+          error: {
+            status: 'PROJECT_NOT_FOUND',
+            message: 'test reason',
+          },
+        };
+        stub = sinon.stub(HttpClient.prototype, 'send');
+        stub.onCall(0).resolves(utils.responseFrom('discovered-service-account'));
+        stub.onCall(1).rejects(utils.errorFrom(expectedResult));
+        const requestHandler = new AuthorizedHttpClient(mockApp);
+        const signer = new IAMSigner(requestHandler);
+        return signer.sign(input).catch((err) => {
+          const message = 'test reason; Please refer to ' +
+            'https://firebase.google.com/docs/auth/admin/create-custom-tokens for more details on ' +
+            'how to use and troubleshoot this feature.';
+          expect(err.message).to.equal(message);
+          expect(stub).to.have.been.calledTwice;
+          expect(stub.getCall(0).args[0]).to.deep.equal(metadataRequest);
+          expect(stub.getCall(1).args[0]).to.deep.equal(signRequest);
+        });
+      });
+
+      it('should return the discovered service account', () => {
+        stub = sinon.stub(HttpClient.prototype, 'send');
+        stub.onCall(0).resolves(utils.responseFrom('discovered-service-account'));
+        const signer = new IAMSigner(new AuthorizedHttpClient(mockApp));
+        return signer.getAccountId().should.eventually.equal('discovered-service-account');
+      });
+
+      it('should return the expected error when failed to contact the Metadata server', () => {
+        stub = sinon.stub(HttpClient.prototype, 'send');
+        stub.onCall(0).rejects(utils.errorFrom('test error'));
+        const signer = new IAMSigner(new AuthorizedHttpClient(mockApp));
+        const expected = 'Failed to determine service account. Make sure to initialize the SDK with ' +
+          'a service account credential. Alternatively specify a service account with ' +
+          'iam.serviceAccounts.signBlob permission.';
+        return signer.getAccountId().should.eventually.be.rejectedWith(expected);
+      });
+    });
+  });
+});
 
 describe('FirebaseTokenGenerator', () => {
   let tokenGenerator: FirebaseTokenGenerator;
 
   let clock: sinon.SinonFakeTimers;
   beforeEach(() => {
-    tokenGenerator = new FirebaseTokenGenerator(new Certificate(mocks.certificateObject));
+    const cert = new Certificate(mocks.certificateObject);
+    tokenGenerator = new FirebaseTokenGenerator(new ServiceAccountSigner(cert));
   });
 
   afterEach(() => {
@@ -86,66 +274,23 @@ describe('FirebaseTokenGenerator', () => {
   });
 
   describe('Constructor', () => {
-    it('should throw given no service account', () => {
+    it('should throw given no arguments', () => {
       expect(() => {
         // Need to overcome the type system to allow a call with no parameter
         const anyFirebaseTokenGenerator: any = FirebaseTokenGenerator;
         return new anyFirebaseTokenGenerator();
-      }).to.throw('Must provide a certificate to use FirebaseTokenGenerator');
+      }).to.throw('Must provide a CryptoSigner to use FirebaseTokenGenerator');
     });
 
-    const invalidCredentials = [null, NaN, 0, 1, true, false, '', 'a', [], {}, { a: 1 }, _.noop];
-    invalidCredentials.forEach((invalidCredential) => {
-      it('should throw given invalid Credential: ' + JSON.stringify(invalidCredential), () => {
+    const invalidSigners = [null, NaN, 0, 1, true, false, '', 'a', [], _.noop];
+    invalidSigners.forEach((invalidSigner) => {
+      it('should throw given invalid signer: ' + JSON.stringify(invalidSigner), () => {
         expect(() => {
-          return new FirebaseTokenGenerator(new Certificate(invalidCredential as any));
-        }).to.throw(Error);
+          return new FirebaseTokenGenerator(invalidSigner as any);
+        }).to.throw('Must provide a CryptoSigner to use FirebaseTokenGenerator');
       });
     });
-
-    it('should throw given an object without a "private_key" property', () => {
-      const invalidCertificate = _.omit(mocks.certificateObject, 'private_key');
-      expect(() => {
-        return new FirebaseTokenGenerator(new Certificate(invalidCertificate as any));
-      }).to.throw('Certificate object must contain a string "private_key" property');
-    });
-
-    it('should throw given an object with an empty string "private_key" property', () => {
-      const invalidCertificate = _.clone(mocks.certificateObject);
-      invalidCertificate.private_key = '';
-      expect(() => {
-        return new FirebaseTokenGenerator(new Certificate(invalidCertificate as any));
-      }).to.throw('Certificate object must contain a string "private_key" property');
-    });
-
-    it('should throw given an object without a "client_email" property', () => {
-      const invalidCertificate = _.omit(mocks.certificateObject, 'client_email');
-      expect(() => {
-        return new FirebaseTokenGenerator(new Certificate(invalidCertificate as any));
-      }).to.throw('Certificate object must contain a string "client_email" property');
-    });
-
-    it('should throw given an object without an empty string "client_email" property', () => {
-      const invalidCertificate = _.clone(mocks.certificateObject);
-      invalidCertificate.client_email = '';
-      expect(() => {
-        return new FirebaseTokenGenerator(new Certificate(invalidCertificate as any));
-      }).to.throw('Certificate object must contain a string "client_email" property');
-    });
-
-    it('should not throw given a valid certificate', () => {
-      expect(() => {
-        return new FirebaseTokenGenerator(new Certificate(mocks.certificateObject));
-      }).not.to.throw();
-    });
-
-    it('should not throw given an object representing a certificate key', () => {
-      expect(() => {
-        return new FirebaseTokenGenerator(mocks.certificateObject);
-      }).not.to.throw();
-    });
   });
-
 
   describe('createCustomToken()', () => {
     it('should throw given no uid', () => {
@@ -202,26 +347,6 @@ describe('FirebaseTokenGenerator', () => {
           tokenGenerator.createCustomToken(mocks.uid, blacklistedDeveloperClaims);
         }).to.throw('Developer claim "' + blacklistedClaim + '" is reserved and cannot be specified');
       });
-    });
-
-    it('should throw if the token generator was initialized with no "private_key"', () => {
-      const certificateObjectWithNoPrivateKey: any = _.omit(mocks.certificateObject, 'private_key');
-      certificateObjectWithNoPrivateKey.clientEmail = certificateObjectWithNoPrivateKey.client_email;
-      const tokenGeneratorWithNoPrivateKey = new FirebaseTokenGenerator(certificateObjectWithNoPrivateKey);
-
-      expect(() => {
-        tokenGeneratorWithNoPrivateKey.createCustomToken(mocks.uid);
-      }).to.throw('createCustomToken() requires a certificate with "private_key" set');
-    });
-
-    it('should throw if the token generator was initialized with no "client_email"', () => {
-      const certificateObjectWithNoClientEmail: any = _.omit(mocks.certificateObject, 'client_email');
-      certificateObjectWithNoClientEmail.privateKey = certificateObjectWithNoClientEmail.private_key;
-      const tokenGeneratorWithNoClientEmail = new FirebaseTokenGenerator(certificateObjectWithNoClientEmail);
-
-      expect(() => {
-        tokenGeneratorWithNoClientEmail.createCustomToken(mocks.uid);
-      }).to.throw('createCustomToken() requires a certificate with "client_email" set');
     });
 
     it('should be fulfilled given a valid uid and no developer claims', () => {
@@ -289,10 +414,9 @@ describe('FirebaseTokenGenerator', () => {
           const decoded: any = jwt.decode(token, {
             complete: true,
           });
-
           expect(decoded.header).to.deep.equal({
-            typ: 'JWT',
             alg: ALGORITHM,
+            typ: 'JWT',
           });
         });
     });
@@ -345,145 +469,4 @@ describe('FirebaseTokenGenerator', () => {
         });
     });
   });
-
-  describe('verifySessionCookie()', () => {
-    const sessionCookie = mocks.generateSessionCookie();
-    const decodedSessionCookie = jwt.decode(sessionCookie);
-    let stubs: sinon.SinonStub[] = [];
-    let tokenVerifierConstructorStub: sinon.SinonStub;
-    let sessionCookieVerifier: any;
-    let idTokenVerifier: any;
-
-    beforeEach(() => {
-      // Create stub instances to be used for session cookie verifier and id token verifier.
-      sessionCookieVerifier = sinon.createStubInstance(verifier.FirebaseTokenVerifier);
-      idTokenVerifier = sinon.createStubInstance(verifier.FirebaseTokenVerifier);
-      // Stub FirebaseTokenVerifier constructor to return stub instance above depending on
-      // issuer.
-      tokenVerifierConstructorStub = sinon.stub(verifier, 'FirebaseTokenVerifier')
-        .callsFake((certUrl, algorithm, issuer, projectId, tokenInfo) => {
-          // Return mock token verifier.
-          if (issuer === 'https://session.firebase.google.com/') {
-            return sessionCookieVerifier;
-          } else {
-            return idTokenVerifier;
-          }
-        });
-      stubs.push(tokenVerifierConstructorStub);
-    });
-
-    after(() => {
-      stubs = [];
-    });
-
-    afterEach(() => {
-      // Confirm token verifiers initialized with expected arguments.
-      expect(tokenVerifierConstructorStub).to.have.been.calledTwice.and.calledWith(
-          'https://www.googleapis.com/identitytoolkit/v3/relyingparty/publicKeys',
-          ALGORITHM,
-          'https://session.firebase.google.com/',
-          'project_id',
-          SESSION_COOKIE_INFO);
-      expect(tokenVerifierConstructorStub).to.have.been.calledTwice.and.calledWith(
-          'https://www.googleapis.com/robot/v1/metadata/x509/securetoken@system.gserviceaccount.com',
-          ALGORITHM,
-          'https://securetoken.google.com/',
-          'project_id',
-          ID_TOKEN_INFO);
-      _.forEach(stubs, (stub) => stub.restore());
-    });
-
-    it('resolves when underlying sessionCookieVerifier.verifyJWT() resolves with expected result', () =>  {
-      sessionCookieVerifier.verifyJWT.withArgs(sessionCookie).returns(Promise.resolve(decodedSessionCookie));
-
-      tokenGenerator = new FirebaseTokenGenerator(new Certificate(mocks.certificateObject));
-
-      return tokenGenerator.verifySessionCookie(sessionCookie)
-        .then((result) => {
-          expect(result).to.deep.equal(decodedSessionCookie);
-        });
-    });
-
-    it('rejects when underlying sessionCookieVerifier.verifyJWT() rejects with expected error', () =>  {
-      const expectedError = new FirebaseAuthError(
-        AuthClientErrorCode.INVALID_ARGUMENT, 'Decoding Firebase session cookie failed');
-      sessionCookieVerifier.verifyJWT.withArgs(sessionCookie).returns(Promise.reject(expectedError));
-
-      tokenGenerator = new FirebaseTokenGenerator(new Certificate(mocks.certificateObject));
-
-      return tokenGenerator.verifySessionCookie(sessionCookie)
-        .should.eventually.be.rejectedWith('Decoding Firebase session cookie failed');
-    });
-  });
-
-  describe('verifyIdToken()', () => {
-    const idToken = mocks.generateIdToken();
-    const decodedIdToken = jwt.decode(idToken);
-    let stubs: sinon.SinonStub[] = [];
-    let tokenVerifierConstructorStub: sinon.SinonStub;
-    let sessionCookieVerifier: any;
-    let idTokenVerifier: any;
-
-    beforeEach(() => {
-      // Create stub instances to be used for session cookie verifier and id token verifier.
-      sessionCookieVerifier = sinon.createStubInstance(verifier.FirebaseTokenVerifier);
-      idTokenVerifier = sinon.createStubInstance(verifier.FirebaseTokenVerifier);
-      // Stub FirebaseTokenVerifier constructor to return stub instance above depending on
-      // issuer.
-      tokenVerifierConstructorStub = sinon.stub(verifier, 'FirebaseTokenVerifier')
-        .callsFake((certUrl, algorithm, issuer, projectId, tokenInfo) => {
-          // Return mock token verifier.
-          if (issuer === 'https://session.firebase.google.com/') {
-            return sessionCookieVerifier;
-          } else {
-            return idTokenVerifier;
-          }
-        });
-      stubs.push(tokenVerifierConstructorStub);
-    });
-
-    after(() => {
-      stubs = [];
-    });
-
-    afterEach(() => {
-      // Confirm token verifiers initialized with expected arguments.
-      expect(tokenVerifierConstructorStub).to.have.been.calledTwice.and.calledWith(
-          'https://www.googleapis.com/identitytoolkit/v3/relyingparty/publicKeys',
-          ALGORITHM,
-          'https://session.firebase.google.com/',
-          'project_id',
-          SESSION_COOKIE_INFO);
-      expect(tokenVerifierConstructorStub).to.have.been.calledTwice.and.calledWith(
-          'https://www.googleapis.com/robot/v1/metadata/x509/securetoken@system.gserviceaccount.com',
-          ALGORITHM,
-          'https://securetoken.google.com/',
-          'project_id',
-          ID_TOKEN_INFO);
-      _.forEach(stubs, (stub) => stub.restore());
-    });
-
-    it('resolves when underlying idTokenVerifier.verifyJWT() resolves with expected result', () =>  {
-      idTokenVerifier.verifyJWT.withArgs(idToken).returns(Promise.resolve(decodedIdToken));
-
-      tokenGenerator = new FirebaseTokenGenerator(new Certificate(mocks.certificateObject));
-
-      return tokenGenerator.verifyIdToken(idToken)
-        .then((result) => {
-          expect(result).to.deep.equal(decodedIdToken);
-        });
-    });
-
-    it('rejects when underlying idTokenVerifier.verifyJWT() rejects with expected error', () =>  {
-      const expectedError = new FirebaseAuthError(
-        AuthClientErrorCode.INVALID_ARGUMENT, 'Decoding Firebase ID token failed');
-      idTokenVerifier.verifyJWT.withArgs(idToken).returns(Promise.reject(expectedError));
-
-      tokenGenerator = new FirebaseTokenGenerator(new Certificate(mocks.certificateObject));
-
-      return tokenGenerator.verifyIdToken(idToken)
-        .should.eventually.be.rejectedWith('Decoding Firebase ID token failed');
-    });
-  });
 });
-